+++ src/constants/apis.js
@@ -20,7 +20,10 @@ export const APIS = {
   HTTP: {
     LCD: LCD,
     API_VERSION: API_VERSION,
-    BACKEND: `${BACKEND}/${BACKEND_VERSION}`,
+    // Backend
+    BACKEND_CHART: `${BACKEND}/${BACKEND_VERSION}/chart`,
+    BACKEND_OVERVIEW: `${BACKEND}/${BACKEND_VERSION}/overview`,
+    BACKEND_VALIDATORS: `${BACKEND}/${BACKEND_VERSION}/validators`,
     // Query
     AUTH: `${LCD}/${API_PATHS.COSMOS}/auth/${API_VERSION}`,
     BANK: `${LCD}/${API_PATHS.COSMOS}/bank/${API_VERSION}`,
@@ -52,8 +55,7 @@ export const APIS = {
   },
   WS: {
     EVENTS: {
-      BACKEND_NEW_BLOCK: 'new-block',
-      BACKEND_NEW_TX: 'new-tx',
+      BACKEND_NEW_HEIGHT: 'NewHeight',
       NEW_BLOCK: 'NewBlock',
       TX: 'Tx',
       TX_SET_PARAMS: 'setParams',