+++ src/store/exchange-rate/getters.js
@@ -1,10 +1,6 @@
 export default {
-  abrTokens: (state) => state.abrTokens,
-  accounts: (state) => state.accounts,
   error: (state) => state.error,
-  freezedTokens: (state) => state.freezedTokens,
   isLoading: (state) => state.isLoading,
-  pool: (state) => state.pool,
-  supply: (state) => state.supply,
-  vbrTokens: (state) => state.vbrTokens,
+  chart: (state) => state.chart,
+  overview: (state) => state.overview,
 };