+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.2.0] - 2022-06-08
+## [alfa] - 3.2.1 - 2022-06-16
+
+## Added
+
+- Add chart backend endpoint
+- Add overview backend endpoint
+
+## Chnaged
+
+- Update fetch validators data from backend
+- Update backend event
+
+## [alfa] - 3.2.0 - 2022-06-08
 
 ### Added
 
@@ -384,7 +396,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add RFC cross chain navigation
 - Add a switch to selcet a dark theme from navigation drawer
 
-[3.2.0]: https://github.com/commercionetwork/almerico/compare/v3.1.6...v3.2.0
 [3.1.6]: https://github.com/commercionetwork/almerico/compare/v3.1.5...v3.1.6
 [3.1.5]: https://github.com/commercionetwork/almerico/compare/v3.1.4...v3.1.5
 [3.1.4]: https://github.com/commercionetwork/almerico/compare/v3.1.3...v3.1.4